// src/util/types/enums.ts
export enum TaskRecurrence {
  Single = 'single',
  Daily = 'daily',
  Weekly = 'weekly',
  Monthly = 'monthly',
  Yearly = 'yearly',
  Cycle = 'cycle',
}

export enum TaskCycleUnit {
  Day = 'day',
  Week = 'week',
  Month = 'month',
}

export enum TaskListType {
  Checklist = 'checklist',
  Rotating = 'rotating',
  Random = 'random',
}